Skeleton change: read from stdin as per lab doc.
NB POSSIBLE AUTOGRADER BREAK

Lab doc told students to use
    ./a.out < testfile.txt
to run their code; however the program actually took in a filename on
stdin.

Modified the code to read its input from stdin. Other option was to
update the lab document, but this is easier & more amenable to
autograding- that said, the autograder is probably not yet set up to use
this input format.