Modified Hash Lab in consideration of #144
I changed getWordInTable to getWordInGrid (along with all other
references to "table" as opposed to "grid" in that cpp file), updated
the getWordInGrid.cpp.html file, and zipped up the new version in code.zip.

I added comments above the getWordInGrid() method to specify that the
output need not be "len" characters long if up against the edge of the
grid.

For Big-Theta, I specified in all areas that we're looking for *only*
the runtime associated with the actual word-finding portion of the
algorithm, and not the whole program.  This was complete reversal of the
postlab text, though--I wasn't entirely sure if this was desired.