Improved/fixed cin code examples
The second cin code example is broken and I have seen issues with it in the past (cin.good() does NOT check if there is any more input). The first version works fine but is verbose and can lead to confusing behavior, as opposed to the normal idiom which is much cleaner and simpler.