Fix memory leak in BST/AVL template
Previously, there had been an indicated memory leak in the
implementations of BinarySearchTree and AVLTree, primarily because we
set parent pointers to NULL, and access previously deleted objects.

This change:
1. Simplifies the students' implementations by providing constructors
and destructors for BinaryNode, BinarySearchTree, AVLNode, and AVLTree.
2. Destructor for Node objects is top-down recursive.
3. Destructor for tree relies on Node destructor behavior.
4. The Remove method will set the left/right pointers to NULL for a Node
with a single child, so that deleting the parent won't recursively
delete the child. The child is still stored in a temp pointer, so we can
safely return it.

The Node classes are now separated into their own header files, along
with their own C++ implementations. The documentation and resources have
been updated accordingly.