Restructure Lab 5
This change restructures lab 5 to incorporate some of the feedback
provided over the Spring 2018 semester. Currently, the post-lab requires
students to complete a binary search tree, avl tree, and report, which
results in odd pacing throughout the typical lab week. Also, it is in
general a good idea to implement the binary search tree before the avl
tree anyways, and this restructuring implicitly reinforces that order.
The in-lab currently only contains a worksheet and Makefile tutorial,
which were previously in the pre-lab section.

This change primarily:
1. Moves the Binary search tree implementation to the in-lab.
2. Returns the Makefile tutorial and AVL worksheet to the pre-lab (we
already know that students are capable of doing this pre-lab).

The restructuring has the following consequences:
1. The Binary search tree and AVL tree are no longer tested together.
Both implementations have separate test runners, under BSTPathTest.cpp
and AVLPathTest.cpp.
2. The Binary search tree and AVL tree have separate test file outputs.
We could have kept them together, having students turn in an
unimplemented AVL tree for the in-lab, but I feel that this would just
complicate instructions. The grading server will have to be changed
accordingly.

The documentation has also been updated accordingly:
1. The bulk of the search tree instructions are now in the in-lab, with
the post-lab listing requirements specific to the AVL implementation.
2. The report lists specific requirements for what it wants in a
discussion. Previously, there was some ambiguity about “specific
numerical results” that was left over from the old post-lab 5.
3. The background section at the top of the lab docs now discusses
search trees as well. Previously it really only talked about expression
trees.